refactor(extensions): move input rules to examples, add doc pages
