+++ .changeset/shaky-ties-give.md
@@ -1,0 +1,6 @@
+---
+"@prosekit/extensions": patch
+"prosekit": patch
+---
+
+add subscript and superscript extension
