+++ packages/extensions/src/subscript/index.ts
@@ -1,4 +1,3 @@
 export { defineSubscriptCommands, type SubscriptCommandsExtension } from './subscript-commands.ts'
-export { defineSubscriptInputRule } from './subscript-input-rule.ts'
 export { defineSubscriptSpec, type SubscriptSpecExtension } from './subscript-spec.ts'
 export { defineSubscript, type SubscriptExtension } from './subscript.ts'
