--- packages/extensions/src/subscript/subscript-input-rule.ts
@@ -1,15 +1,0 @@
-import { canUseRegexLookbehind, type PlainExtension } from '@prosekit/core'
-
-import { defineMarkInputRule } from '../input-rule/index.ts'
-
-/**
- * @internal
- */
-export function defineSubscriptInputRule(): PlainExtension {
-  return defineMarkInputRule({
-    regex: canUseRegexLookbehind()
-      ? /(?<=\s|^)~([^\s~]|[^\s~][^~]*[^\s~])~$/
-      : /~([^\s~]|[^\s~][^~]*[^\s~])~$/,
-    type: 'subscript',
-  })
-}
