+++ packages/extensions/src/subscript/subscript.ts
@@ -1,7 +1,6 @@
 import { union, type Union } from '@prosekit/core'
 
 import { defineSubscriptCommands, type SubscriptCommandsExtension } from './subscript-commands.ts'
-import { defineSubscriptInputRule } from './subscript-input-rule.ts'
 import { defineSubscriptSpec, type SubscriptSpecExtension } from './subscript-spec.ts'
 
 /**
@@ -15,6 +14,5 @@ export function defineSubscript(): SubscriptExtension {
   return union(
     defineSubscriptSpec(),
     defineSubscriptCommands(),
-    defineSubscriptInputRule(),
   )
 }
