+++ packages/extensions/src/superscript/index.ts
@@ -1,4 +1,3 @@
 export { defineSuperscriptCommands, type SuperscriptCommandsExtension } from './superscript-commands.ts'
-export { defineSuperscriptInputRule } from './superscript-input-rule.ts'
 export { defineSuperscriptSpec, type SuperscriptSpecExtension } from './superscript-spec.ts'
 export { defineSuperscript, type SuperscriptExtension } from './superscript.ts'
