+++ packages/extensions/src/superscript/superscript.ts
@@ -1,7 +1,6 @@
 import { union, type Union } from '@prosekit/core'
 
 import { defineSuperscriptCommands, type SuperscriptCommandsExtension } from './superscript-commands.ts'
-import { defineSuperscriptInputRule } from './superscript-input-rule.ts'
 import { defineSuperscriptSpec, type SuperscriptSpecExtension } from './superscript-spec.ts'
 
 /**
@@ -15,6 +14,5 @@ export function defineSuperscript(): SuperscriptExtension {
   return union(
     defineSuperscriptSpec(),
     defineSuperscriptCommands(),
-    defineSuperscriptInputRule(),
   )
 }
