+++ website/src/content/docs/extensions/superscript.mdx
@@ -1,0 +1,50 @@
+---
+title: Superscript
+description: "Superscript text formatting backed by the standard <sup> mark"
+---
+
+import Demo from '@/components/ui/demo/demo.astro';
+
+The `superscript` mark is used to represent text that is superscripted (raised above the baseline). It will be rendered as `<sup>` element in HTML.
+
+<Demo story="sub-sup" />
+
+## Commands
+
+### `toggleSuperscript`
+
+```ts twoslash
+import { createEditor } from 'prosekit/core'
+import { defineSuperscript } from 'prosekit/extensions/superscript'
+
+const extension = defineSuperscript()
+const editor = createEditor({ extension })
+// ---cut---
+editor.commands.toggleSuperscript()
+```
+
+## Input Rules
+
+Input rules are not included by default. You can add them manually using `defineMarkInputRule` from `prosekit/extensions/input-rule`:
+
+```ts twoslash
+import { canUseRegexLookbehind, union } from 'prosekit/core'
+import { defineMarkInputRule } from 'prosekit/extensions/input-rule'
+import { defineSuperscript } from 'prosekit/extensions/superscript'
+
+const extension = union(
+  defineSuperscript(),
+  defineMarkInputRule({
+    regex: canUseRegexLookbehind()
+      ? /(?<=\s|^)\^([^\s^]|[^\s^][^^]*[^\s^])\^$/
+      : /\^([^\s^]|[^\s^][^^]*[^\s^])\^$/,
+    type: 'superscript',
+  }),
+)
+```
+
+Typing `^text^` followed by a space will automatically convert the text to superscript.
+
+## API Reference
+
+- [prosekit/extensions/superscript](/references/extensions/superscript)
