+++ website/src/sidebar/extensions.json
@@ -26,6 +26,8 @@
       "italic",
       "link",
       "strike",
+      "subscript",
+      "superscript",
       "text-color",
       "underline"
     ]
